# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Cellphones API",
        "description": "API documentation for the Cellphones system",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Main API server"
        }
    ],
    "paths": {
        "/api/ping": {
            "get": {
                "tags": [
                    "Health Check"
                ],
                "summary": "Health check endpoint",
                "description": "Returns a simple pong message to verify that the API is running.",
                "operationId": "ec46de3798425d731e23605b41ebe2c2",
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "pong"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/auth/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Đăng ký người dùng",
                "operationId": "ee65fb3f9c9eb5626a80236e13c1683d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RegisterRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Đăng ký thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Đăng ký thành công"
                                        },
                                        "token": {
                                            "type": "string",
                                            "example": "jwt.token.here"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/User"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "RegisterRequest": {
                "required": [
                    "name",
                    "email",
                    "password"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Nguyen Van A"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "a@example.com"
                    },
                    "password": {
                        "type": "string",
                        "format": "password",
                        "example": "123456"
                    }
                },
                "type": "object"
            },
            "User": {
                "title": "User Resource",
                "required": [
                    "id",
                    "name",
                    "email"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "uuid": {
                        "type": "string",
                        "example": "123e4567-e89b-12d3-a456-426614174000"
                    },
                    "name": {
                        "type": "string",
                        "example": "Nguyen Van A"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "a@example.com"
                    },
                    "role_id": {
                        "type": "integer",
                        "example": 2
                    },
                    "image_url": {
                        "type": "string",
                        "format": "uri",
                        "example": "http://localhost/storage/avatar.jpg"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-06-09T10:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-06-09T10:00:00Z"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Nhập token JWT vào ô Authorization",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Auth",
            "description": "Quản lý xác thực người dùng"
        },
        {
            "name": "Health Check",
            "description": "Health Check"
        }
    ]
}